Refactored Collection to interact more flexibly with Tracer.
Collections now specify the amount of data that ray states occupy,
for interrupting and resuming traversal.  This provides more
flexibility and a less arbitrary interface to traversing collections,
since different collections may require different amounts of data
to store ray state.

Collections also now report the number of primitives they contain.
Combined with more compact id's (not yet implemented), this will
allow for much more efficient potential intersection sorting in
the future.

Lastly, implemented an alternative Vec3 class that uses sse intrinsics.
It doesn't improve performance at the moment, however, so I've disabled
it.  But it will probably be useful once we start doing a lot of matrix
transforms down the road, at which point matrix multiplication can
also be implemented with intrinsics.

+++ tracer/tracer.hpp
@@ -45,7 +45,7 @@ public:
 	int thread_count;
 
 	Array<RayInter *> ray_inters; // Ray/Intersection data
-	Array<uint64> states; // Ray states, for interrupting and resuming traversal
+	Array<byte> states; // Ray states, for interrupting and resuming traversal
 	Array<PotentialInter> potential_inters; // "Potential intersection" buffer
 
 	Tracer(Scene *scene_, int thread_count_=1) {
